// makerchecker/src/app/model/users.ts
export interface Customer {
  customer_id: number;
  customer_code: string;
  customer_name: string;
  customer_address1: string;
  customer_address2: string;
  customer_pincode: number;
  email: string;
  contact_number: number;
  primary_contact_person: string;
  record_status: string;
  is_enabled: string;
}